pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llama-finetune"
version = "0.1.0"
description = "Fine-tuning pipeline for LLaMA using Unsloth and Hugging Face"
authors = [
  { name = "Your Name", email = "you@example.com" }
]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
  "unsloth>=2025.4.7",
  "transformers>=4.51.3",
  "datasets>=2.19.0",
  "accelerate>=0.34.1",
  "trl>=0.15.2",
  "pydantic>=2.0",
  "pyyaml>=6.0",
  "rich>=13.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "black>=24.0",
  "ruff>=0.4",
  "mypy>=1.8"
]

docs = [
  "mkdocs>=1.5",
  "mkdocs-material>=9.5"
]

train = [
  "scipy>=1.11",
  "scikit-learn>=1.4"
]

[tool.setuptools.packages.find]
where = ["src"]
